Progress on building Tableau class
